# Provide aliasing hint to the compiler